+++ src/app.scss
@@ -281,7 +281,6 @@ a{
     align-items: center;
     justify-content: space-between;
     width: 100%;
-    border-bottom: 1px solid #dfdfdf;
     padding: 0 1.2rem;
   }
 }
